component folder stucture refactor clean
