+++ src/components/LoadingWeb3.js
@@ -3,8 +3,8 @@ import { makeStyles } from '@material-ui/core/styles';
 
 import Typography from '@material-ui/core/Typography';
 import LinearProgress from '@material-ui/core/LinearProgress';
-import Header from '../Header/Header';
-import EllipsisLoader from '../EllipsisLoader/EllipsisLoader';
+import Header from './Header';
+import EllipsisLoader from './EllipsisLoader';
 const useStyles = makeStyles({
   root: {
     width: '100%',
