--- src/components/ScrollReset/index.js
@@ -1,1 +1,0 @@
-export { default } from './ScrollReset';
