+++ src/views/SubmitProposal.js
@@ -8,10 +8,10 @@ import { useForm } from 'react-hook-form';
 import useRouter from '../utils/useRouter';
 import { pinHash } from '../modules/pinata';
 import { Page } from '../components';
-import Header from '../components/Header/Header';
+import Header from '../components/Header';
 import useWeb3Connect from '../utils/useWeb3Connect';
 import HowToVoteIcon from '@material-ui/icons/HowToVote';
-import LoadingWeb3 from '../components/LoadingWeb3/LoadingWeb3';
+import LoadingWeb3 from '../components/LoadingWeb3';
 import Stepper from '@material-ui/core/Stepper';
 import Step from '@material-ui/core/Step';
 import StepLabel from '@material-ui/core/StepLabel';
@@ -38,7 +38,7 @@ import 'react-mde/lib/styles/css/react-mde-all.css';
 import { emojiExists } from '../modules/twitterDb';
 import IpfsUpload from '../components/IpfsUpload';
 // import Box from '3box';
-import ProposalCard from '../components/ProposalCard/ProposalCard';
+import ProposalCard from '../components/ProposalCard';
 import { useRedirectHomeIfNoEthAccount } from '../utils/useCommonUtils';
 
 const BN = require('bn.js');
